only one bot file was left
